+++ packages/colony-js-contract-client/src/defaults.js
@@ -1,7 +1,0 @@
-/* @flow */
-
-// eslint-disable-next-line import/prefer-default-export
-export const DEFAULT_SEND_OPTIONS = {
-  timeoutMs: 1000 * 60 * 3, // 3 minutes
-  waitForMining: true,
-};